[FIX+IMP] helpdesk_mgmt: Use standard user assignation
Odoo already includes a mechanism for notifying users when a document
has been assigned to them, through the convention of the field name
(`user_id`) and `tracking=True`.

Having to perform this manually is a lot of code to maintain, but it's
also right now incorrect, as the notification is being sent also to the
customer.

We switch here to the standard system.

An index is also good over this field that is widely used.

TT36502